Restructuring tests This puts tests in different folders
